refactor(portfolio): 필요없는 코드 수정
