+++ apps/portfolio/src/entities/project/index.ts
@@ -1,18 +1,5 @@
 export * from "./model/types";
-export * from "./model/service";
+export * from "./model/repository";
 export * from "./model/queries";
 export * from "./model/catalog";
 export * from "./utils";
-
-// Re-export utils for convenience
-export {
-  getThumbnailPath,
-  getThumbnailAspectRatio,
-  getFallbackThumbnail,
-} from "../../utils/thumbnail";
-export {
-  formatProjectName,
-  normalizeTags,
-  normalizeBoolean,
-  normalizeNumber,
-} from "../../utils/normalize";
