+++ apps/portfolio/src/entities/project/model/queries.ts
@@ -1,5 +1,5 @@
 import { useMemo } from "react";
-import { getPortfolioIndex, getProject } from "./service";
+import { getPortfolioIndex, getProject } from "./repository";
 import type { ProjectIndex } from "./types";
 
 export function usePortfolioIndex(): ProjectIndex | undefined {
